bootstrap: fix ws and tabs

+++ bootstrap.sh
@@ -71,36 +71,36 @@ doStow() {
       shopt -s dotglob
       for dotfile in ${1}/.* ; do
         if [ ! -f "${dotfile}" ]; then
-	  #
-	  # Skip the . and .. files as they don't count
-	  #
-	  if [[ ("${dotfile}" == "${1}/." ) || ( "${dotfile}" == "${1}/.." ) ]]; then
-	    debug "Skipping"
+          #
+          # Skip the . and .. files as they don't count
+          #
+          if [[ ("${dotfile}" == "${1}/." ) || ( "${dotfile}" == "${1}/.." ) ]]; then
+            debug "Skipping"
             continue
-	  fi
+          fi
 
-	  #
-	  # Work around to deal with a whole .dir
-	  #
+          #
+          # Work around to deal with a whole .dir
+          #
           ln -s "${dotfile}" "${HOME}/${dotfile##*/}"
         else
           debug "We have a dotfile: ${dotfile}"
-	  orig_file=$(basename "${dotfile}")
+          orig_file=$(basename "${dotfile}")
 
-	  #
-	  # Check for a version in the $HOME
-	  # If there is one, delete it
-	  #
-	  if [ -f "${HOME}/${orig_file}" ]; then
-            debug "Removing original file"
-	    rm -Rf "${HOME}/${orig_file}"
-	  fi
-	  if [ -L "${HOME}/${orig_file}" ]; then
-            debug "Removing soft linked file"
-	    rm -Rf "${HOME}/${orig_file}"
-	  fi
+          #
+          # Check for a version in the $HOME
+          # If there is one, delete it
+          #
+          if [ -f "${HOME}/${orig_file}" ]; then
+                debug "Removing original file"
+            rm -Rf "${HOME:?}/${orig_file}"
+          fi
+          if [ -L "${HOME}/${orig_file}" ]; then
+                debug "Removing soft linked file"
+            rm -Rf "${HOME:?}/${orig_file}"
+          fi
 
-	  debug "Linking ${HOME}/${dotfile##*/} with ${DOTFILES}/${dotfile}"
+          debug "Linking ${HOME}/${dotfile##*/} with ${DOTFILES}/${dotfile}"
           ln -s "${DOTFILES}/${dotfile}" "${HOME}/${dotfile##*/}"
         fi
       done
